feat: introduce suite-sync-types to handle circular imports in extraDeendencies
