+++ suite-common/redux-utils/tsconfig.json
@@ -4,7 +4,7 @@
     "references": [
         { "path": "../metadata-types" },
         { "path": "../secure-storage" },
-        { "path": "../suite-sync-storage" },
+        { "path": "../suite-sync-types" },
         { "path": "../suite-types" },
         { "path": "../wallet-config" },
         { "path": "../wallet-types" },
