+++ suite-common/suite-sync-types/.depcheckrc.json
@@ -1,0 +1,7 @@
+{
+    "ignore-patterns": ["libDev", "lib"],
+    "ignores": [
+        "// Todo: This need so be solved. See: https://github.com/trezor/trezor-suite/issues/21552",
+        "@suite-common/wallet-core"
+    ]
+}
