+++ suite-common/suite-sync-types/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/test-utils",
+    "name": "@suite-common/suite-sync-types",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
@@ -10,22 +10,12 @@
         "type-check": "yarn g:tsc --build"
     },
     "dependencies": {
-        "@mobily/ts-belt": "^3.13.1",
         "@reduxjs/toolkit": "2.10.1",
-        "@suite-common/redux-utils": "workspace:*",
-        "@suite-common/secure-storage": "workspace:*",
         "@suite-common/suite-sync-storage": "workspace:*",
-        "@suite-common/suite-sync-types": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
         "@trezor/connect": "workspace:*",
-        "@trezor/device-utils": "workspace:*",
-        "@trezor/type-utils": "workspace:*",
-        "@trezor/utils": "workspace:*",
-        "core-js": "^3.46.0",
-        "fake-indexeddb": "^6.2.4",
-        "redux": "^5.0.1",
-        "redux-thunk": "^3.1.0"
+        "@trezor/type-utils": "workspace:*"
     }
 }
