+++ suite-common/suite-sync-types/src/labeling/updateAccountLabel.ts
@@ -1,0 +1,13 @@
+import { SuiteSyncStorageRepositoryDep } from '../SuiteSyncStorageRepository';
+
+export type UpdateAccountLabelParams = {
+    deviceStaticSessionId: string;
+    accountKey: string;
+    label: string | null;
+};
+
+export type UpdateAccountLabel = (params: UpdateAccountLabelParams) => void;
+
+export type UpdateAccountLabelDeps = { getState: () => any } & SuiteSyncStorageRepositoryDep;
+
+export type UpdateAccountLabelDep = { updateAccountLabel: UpdateAccountLabel };
