+++ suite-common/suite-sync-types/src/turnOnSuteSync.ts
@@ -1,0 +1,10 @@
+import { Dispatch } from '@reduxjs/toolkit';
+
+export type TurnOnSuiteSync = () => void;
+
+export type TurnOnSuiteSyncDep = { turnOnSuiteSync: TurnOnSuiteSync };
+
+export type CreateTurnOnSuiteSyncDeps = {
+    getState: () => any;
+    dispatch: Dispatch;
+};
