+++ suite-common/suite-sync/package.json
@@ -15,6 +15,7 @@
         "@suite-common/metadata-types": "workspace:*",
         "@suite-common/secure-storage": "workspace:*",
         "@suite-common/suite-sync-storage": "workspace:*",
+        "@suite-common/suite-sync-types": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/suite-utils": "workspace:*",
         "@suite-common/test-utils": "workspace:*",
