+++ suite-common/suite-sync/src/SuiteSyncStorageRepository.ts
@@ -1,21 +1,9 @@
-import { CreateSuiteStorageDep, SuiteSyncStorage } from '@suite-common/suite-sync-storage';
+import { SuiteSyncStorage } from '@suite-common/suite-sync-storage';
+import {
+    CreateSuiteSyncStorageRepositoryFactoryDeps,
+    SuiteSyncStorageRepository,
+} from '@suite-common/suite-sync-types';
 import { SuiteSyncOwner, SuiteSyncOwnerId } from '@suite-common/suite-types';
-
-export type CreateSuiteSyncStorageRepositoryFactoryDeps = CreateSuiteStorageDep & {
-    defaultRelayUrl: string;
-    getRelayUrl: () => string | null;
-};
-
-export type SuiteSyncStorageRepository = {
-    get: (owner: SuiteSyncOwner) => SuiteSyncStorage;
-    delete: (ownerId: SuiteSyncOwnerId) => Promise<void>;
-};
-
-export type CreateSuiteSyncStorageRepository = () => SuiteSyncStorageRepository;
-
-export type SuiteSyncStorageRepositoryDep = {
-    suiteSyncStorageRepository: SuiteSyncStorageRepository;
-};
 
 /**
  * Every Wallet has its own SuiteSyncStorage with Owner derived
