+++ suite-common/suite-sync/src/device/ensureSuiteSyncOwnerKeys.ts
@@ -1,48 +1,13 @@
-import { CreateSuiteSyncOwner, CreateSuiteSyncOwnerError } from '@suite-common/suite-sync-storage';
-import {
-    DelegatedIdentityKey,
-    SuiteSyncOwner,
-    TrezorDeviceWithState,
-} from '@suite-common/suite-types';
+import { EnsureSuiteSyncOwnerDeps, EnsureSuiteSyncOwnerKeys } from '@suite-common/suite-sync-types';
 import {
     DeviceCancelledErr,
     DeviceError,
-    ProofOfDelegatedSignFailed,
     getProofOfDelegatedIdentity,
     isCanceledErrorMessage,
 } from '@suite-common/wallet-core';
-import {
-    ProofOfDelegatedSignFailed,
-    getProofOfDelegatedIdentity,
-} from '@suite-common/wallet-core/src/device/delegatedIdentityKey/getProofOfDelegatedIdentity';
-// Todo: move to device utils
-import TrezorConnect from '@trezor/connect';
-import { Result, err } from '@trezor/type-utils';
+import { err } from '@trezor/type-utils';
 
 const PROOF_OF_DELEGATED_IDENTITY_HEADER = 'EvoluGetNode';
-
-export type EnsureSuiteSyncOwnerDeps = {
-    createSuiteSyncOwner: CreateSuiteSyncOwner;
-    trezorConnect: Pick<typeof TrezorConnect, 'evoluGetNode'>;
-};
-
-export type EnsureSuiteSyncOwnerKeysParams = {
-    device: Pick<TrezorDeviceWithState, 'useEmptyPassphrase' | 'path' | 'state' | 'instance'>;
-    delegatedKey: DelegatedIdentityKey;
-};
-
-export type EnsureSuiteSyncOwnerKeys = (
-    params: EnsureSuiteSyncOwnerKeysParams,
-) => Promise<
-    Result<
-        SuiteSyncOwner,
-        DeviceCancelledErr | DeviceError | ProofOfDelegatedSignFailed | CreateSuiteSyncOwnerError
-    >
->;
-
-export type EnsureSuiteSyncOwnerDep = {
-    ensureSuiteSyncOwnerKeys: EnsureSuiteSyncOwnerKeys;
-};
 
 export const createEnsureSuiteSyncOwnerKeys =
     (deps: EnsureSuiteSyncOwnerDeps): EnsureSuiteSyncOwnerKeys =>
