+++ suite-common/suite-sync/src/refreshSuiteSyncKeys.ts
@@ -1,54 +1,11 @@
-import { Dispatch } from '@reduxjs/toolkit';
-
-import { CreateSuiteSyncOwnerError } from '@suite-common/suite-sync-storage';
-import { TrezorDevice } from '@suite-common/suite-types';
 import {
-    DeviceCancelledErr,
-    DeviceError,
-    EnsureDelegatedIdentityKeyDep,
-    ProofOfDelegatedSignFailed,
-    deviceActions,
-    selectDevices,
-} from '@suite-common/wallet-core';
+    RefreshSuiteKeysUnavailable,
+    RefreshSuiteSyncKeys,
+    RefreshSuiteSyncKeysDeps,
+} from '@suite-common/suite-sync-types';
+import { deviceActions, selectDevices } from '@suite-common/wallet-core';
 import { isTrezorDeviceWithState } from '@suite-common/wallet-utils';
-import { Result, err, ok } from '@trezor/type-utils';
-
-import { EnsureSuiteSyncOwnerDep } from './device/ensureSuiteSyncOwnerKeys';
-
-export type RefreshSuiteSyncKeysDeps = {
-    getState: () => any;
-    dispatch: Dispatch;
-} & EnsureSuiteSyncOwnerDep &
-    EnsureDelegatedIdentityKeyDep;
-
-type RefreshSuiteSyncKeysParams = {
-    device: TrezorDevice;
-};
-
-export type RefreshSuiteKeysUnavailable = {
-    type: 'RefreshSuiteKeysUnavailable';
-};
-
-/**
- * Device is not connected or device is in a state/configuration, that does not
- * support Suite Sync.
- */
-export const RefreshSuiteKeysUnavailable = (): RefreshSuiteKeysUnavailable => ({
-    type: 'RefreshSuiteKeysUnavailable',
-});
-
-export type RefreshSuiteSyncKeys = (
-    params: RefreshSuiteSyncKeysParams,
-) => Promise<
-    Result<
-        void,
-        | DeviceError
-        | DeviceCancelledErr
-        | RefreshSuiteKeysUnavailable
-        | ProofOfDelegatedSignFailed
-        | CreateSuiteSyncOwnerError
-    >
->;
+import { err, ok } from '@trezor/type-utils';
 
 export const createRefreshSuiteSyncKeys =
     (deps: RefreshSuiteSyncKeysDeps): RefreshSuiteSyncKeys =>
