+++ suite-common/suite-sync/src/relay/changeRelayUrl.ts
@@ -1,21 +1,7 @@
-import { Dispatch } from '@reduxjs/toolkit';
-
-import { SuiteSyncOwner } from '@suite-common/suite-types';
+import { ChangeRelayUrl, ChangeRelayUrlDeps } from '@suite-common/suite-sync-types';
 
 import { setSuiteSyncRelayUrl } from '../suiteSyncActions';
 import { DEFAULT_SUITE_SYNC_RELAY_URL } from './relayUrl';
-import { SuiteSyncStorageRepositoryDep } from '../SuiteSyncStorageRepository';
-
-export type ChangeRelayUrlDeps = {
-    getAllDevicesOwners: () => SuiteSyncOwner[];
-    dispatch: Dispatch;
-} & SuiteSyncStorageRepositoryDep;
-
-export type ChangeRelayUrl = (params: { relayUrl: string | null }) => Promise<void>;
-
-export type ChangeRelayUrlDep = {
-    changeRelayUrl: ChangeRelayUrl;
-};
 
 export const createChangeRelayUrl =
     (deps: ChangeRelayUrlDeps): ChangeRelayUrl =>
