+++ suite-common/suite-sync/src/relay/tests/changeRelayUrl.test.ts
@@ -1,10 +1,11 @@
 import type { UnknownAction } from '@reduxjs/toolkit';
 
 import { SuiteSyncStorage } from '@suite-common/suite-sync-storage';
+import { ChangeRelayUrlDeps } from '@suite-common/suite-sync-types';
 import { asSuiteSyncOwnerId, asSuiteSyncOwnerSecretHex } from '@suite-common/suite-types';
 
 import { mockNotExpected } from '../../../tests/utils';
-import { ChangeRelayUrlDeps, createChangeRelayUrl } from '../changeRelayUrl';
+import { createChangeRelayUrl } from '../changeRelayUrl';
 
 const owner1 = {
     ownerId: asSuiteSyncOwnerId('OwnerId_1'),
