+++ suite-common/suite-sync/src/storage/subscriptionStorage.ts
@@ -1,26 +1,6 @@
+import { SubscriptionName, SubscriptionStorage } from '@suite-common/suite-sync-types';
 import { SuiteSyncOwnerId } from '@suite-common/suite-types';
 import { typedObjectValues } from '@trezor/utils';
-
-/**
- * SuiteSync can handle multiple domains.
- * Currently, we have "labeling" only.
- */
-type SubscriptionName = 'labeling';
-
-type AddSubscriptionParams = {
-    ownerId: SuiteSyncOwnerId;
-    unsubscribe: () => void;
-    name: SubscriptionName;
-};
-
-export type SubscriptionStorage = {
-    add: (params: AddSubscriptionParams) => void;
-    disposeAll: (ownerId: SuiteSyncOwnerId) => void;
-};
-
-export type SubscriptionStorageDep = {
-    subscriptionStorage: SubscriptionStorage;
-};
 
 export const createSubscriptionStorage = (): SubscriptionStorage => {
     const storage: Record<SuiteSyncOwnerId, Partial<Record<SubscriptionName, () => void>>> = {};
