+++ suite-common/suite-sync/tsconfig.json
@@ -6,9 +6,9 @@
     },
     "references": [
         { "path": "../metadata-types" },
-        { "path": "../redux-utils" },
         { "path": "../secure-storage" },
         { "path": "../suite-sync-storage" },
+        { "path": "../suite-sync-types" },
         { "path": "../suite-types" },
         { "path": "../suite-utils" },
         { "path": "../test-utils" },
