+++ suite-common/test-utils/tsconfig.json
@@ -5,6 +5,7 @@
         { "path": "../redux-utils" },
         { "path": "../secure-storage" },
         { "path": "../suite-sync-storage" },
+        { "path": "../suite-sync-types" },
         { "path": "../suite-types" },
         { "path": "../wallet-config" },
         { "path": "../wallet-types" },
