+++ suite-native/labeling/package.json
@@ -12,6 +12,7 @@
     "dependencies": {
         "@gorhom/bottom-sheet": "5.2.6",
         "@reduxjs/toolkit": "2.10.1",
+        "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/validators": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
