+++ suite-native/suite-sync/package.json
@@ -16,7 +16,8 @@
         "@suite-common/secure-storage": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-sync-evolu": "workspace:*",
-        "@suite-common/suite-sync-storage": "workspace:*",
+        "@suite-common/suite-sync-types": "workspace:*",
+        "@suite-common/wallet-core": "workspace:*",
         "@trezor/connect": "workspace:*",
         "expo": "~54.0.20",
         "expo-sqlite": "^16.0.8",
