+++ suite-native/suite-sync/src/createSuiteSyncNativeCompositionRoot.ts
@@ -2,13 +2,14 @@ import { evoluReactNativeDeps } from '@evolu/react-native/expo-sqlite';
 import { Dispatch } from '@reduxjs/toolkit';
 
 import { SecureStorageDep } from '@suite-common/secure-storage';
-import { SuiteSync, createSuiteSyncCompositionRoot } from '@suite-common/suite-sync';
+import { createSuiteSyncCompositionRoot } from '@suite-common/suite-sync';
 import {
     createEvoluInstanceFactory,
     createEvoluStorageFactory,
     evoluCreateSuiteSyncOwner,
 } from '@suite-common/suite-sync-evolu';
-import { EnsureDelegatedIdentityKeyDep , selectAllDeviceOwners } from '@suite-common/wallet-core';
+import { SuiteSync } from '@suite-common/suite-sync-types';
+import { EnsureDelegatedIdentityKeyDep } from '@suite-common/wallet-core';
 import { TrezorConnect } from '@trezor/connect';
 
 type InitSuiteSyncNativeDeps = {
@@ -26,6 +27,5 @@ export const createSuiteSyncNativeCompositionRoot = (deps: InitSuiteSyncNativeDe
         ...deps,
         createSuiteStorage: createEvoluStorage,
         createSuiteSyncOwner: evoluCreateSuiteSyncOwner,
-        getAllDevicesOwners: () => selectAllDeviceOwners(deps.getState()),
     });
 };
