+++ suite/suite-sync/package.json
@@ -15,7 +15,8 @@
         "@suite-common/secure-storage": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-sync-evolu": "workspace:*",
-        "@suite-common/suite-sync-storage": "workspace:*",
+        "@suite-common/suite-sync-types": "workspace:*",
+        "@suite-common/wallet-core": "workspace:*",
         "@trezor/connect": "workspace:*"
     }
 }
